[Refactor] [FE-12] ContentScreen에서 ContentMain 분리
